refact: factor out SearchForm component from Search
